requirements.txt
requests
beautifulsoup4
aiogram
python-dotenv